adding follow me as an example use of the HAL
